+++ src/main/java/lotto/domain/PurchaseInfo.java
@@ -1,6 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.lottoticket.LottoTicket;
+import lotto.domain.ticket.LottoTicket;
 
 public class PurchaseInfo {
     public static final int MANUAL_COUNT_LIMIT = 1;
